requirements.txt
#Langchain Core
langchain
langchain-core

# OpenAI Intergration
langchain-openai
openai

# Anthropic Intergration
langchain-anthropic

# Google Gemini (PaLM) Intergration
langchain-google-genai
google-generativeai

# Hugging Face Intergration
langchain-huggingface
transformers
huggingface-hub

#Envioronment Variable Managment
python-dotenv
streamlit

#Machine Learning Utilities
numpy
scikit-learn
typing-extensions